+++ lib/spider.rb
@@ -1,22 +1,16 @@
 module Whitewidow
   def spider
     puts <<-_END_
-
       ( W ( h ( i ( t ( e ( w ( i ( d ( o ( w )
                     ( S ( Q ( L )
                   ( V ( u ( l ( n )
             ( S ( c ( a ( n ( n ( e ( r )
-
                       _\\{"}/_
                        //^\\\\
-
     _END_
   end
 
   def version
-    '1.0.5'
+    '1.0.6'
   end
 end
-
-
-