+++ docs/CHANGELOG.md
@@ -1,3 +1,20 @@
+#Version 1.0.6
+- Finished optparser, read the readme your options have changed!
+- Edited version numbers
+- Added new contributor
+- Edited file formatter
+- Updated legal disclaimer and term of service.
+- Added three new sql vuln syntax parameters. ( ` ; -- ) all found vuln sites using the default setting will be saved
+with all of the sql syntax params and tested.
+- File is automatically formatted now and will be run as `#sites.txt`
+- Automatic file formatting added
+- Added more information when site is found including, server info and IP address
+- Complete refractor of when site is found, see screenshots for examples
+
+#Version 1.0.5 Patches (1.0.5.1/1.0.5.2)
+- Fixed copy module parameters
+- Patched RestClient error
+
 #Version 1.0.5
  - Added docs dir, this has the copyright info, changelog, and credits to contributors in it.
  - Edited credits removed contributors and put them in the docs directory.
@@ -10,6 +27,7 @@ to push enter to continue.
  - Added simple file formatter in lib/extra for when you have blank line in your file. For this to work move the file
 into the extra directory and run the file as an ARGV against file_formatter.rb. The file wil be saved as `#sites.txt`
 Example: `ruby file_formatter.rb sites.txt`
+
 #Version 1.0.4
  - Started working on optparser for parsing ARGV arguments instead of using ARGV. Optparser will not only be more efficient
  but will make whitewidow more readable and will be more user friendly. The new options will be released upon release of